fix many reference related errors and setup initial kernel test
this does a single pass unite to a second buffer, and looks like it is
working as intended.
reference errors are based on the fact that the struct should not be
passed to the global kernel as a reference or pointer, since it is a
host side construct that simply contains a copy of the device side
pointer (this was causing many issues).
also reworked the .at() to just use .set() and .get() since we can't
really throw when inside the kernel, so it's better to no-op when out of
bounds.