encapsulate most unionfind logic required into subclass DeviceIndexImage
this struct which inherits from DeviceImage handles the pairwise types
like short2 and has most of the logic refactored into it, like
making/checking invalid row/col indices for inactive sites, accounting
for global offset of the tile etc. the logic for unite() and find() now
explicitly use this struct type, so it is slightly clearer imo. the
entire shared mem logic is now calculated using this struct type which
uses the pair row/col indices, as this is required to more easily
account for the global row/col offset when writing back to global mem
(instead of flattening to 1D and then having to perform divisions to go
back to row/col so we can offset it)