compare with naive equivalent If kernel
as usual, only on large numbers will the timing be worse e.g. at 1M
elements then it's roughly twice the duration of the cub's two kernel
calls (including overhead gaps). but when we increase it to the BigData
struct instead of raw ints, the memcpy bandwidth is the bottleneck, and
the two are essentially equivalent. hence if dealing with large structs,
probably can just always do the naive way.